requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
PyJWT==2.9.0
pydantic==2.9.2
httpx==0.27.2
